Refactor run-source attribution utilities
