+++ framework/py/flwr/superlink/extensions.py
@@ -20,42 +20,20 @@ from copy import deepcopy
 from importlib import import_module
 from logging import WARNING
 from types import ModuleType
-from typing import Any, Literal, cast, get_args
+from typing import Any, Literal, cast
 
 from fastapi import FastAPI
 from starlette.middleware import Middleware
 
 from flwr.common.logger import log
 from flwr.supercore.run import Run
+from flwr.superlink.run_source import RunStartSource
 
 SuperLinkLifespanContext = Callable[
     [FastAPI], AbstractAsyncContextManager[Mapping[str, Any] | None]
 ]
-RunStartSource = Literal["cli", "web_ui", "automation", "unknown"]
 ResultDeliveryChannel = Literal["logs", "chat"]
-RUN_SOURCE_METADATA_KEY = "x-flwr-run-source"
-_RUN_START_SOURCES = frozenset(get_args(RunStartSource))
 _SGXT_MODULE = "flwr.ee.superlink.extensions"
-
-
-def resolve_run_start_source(value: str | bytes | None) -> RunStartSource:
-    """Normalize a caller-provided source label for analytics.
-
-    Source attribution is intentionally best effort. Callers can only affect
-    the analytics label for their own request, so recognized values are
-    trusted and invalid values fall back to ``unknown``. This value is not a
-    security or authorization signal.
-    """
-    if value is None:
-        return "unknown"
-    if isinstance(value, bytes):
-        try:
-            value = value.decode("ascii")
-        except UnicodeDecodeError:
-            return "unknown"
-    if value not in _RUN_START_SOURCES:
-        return "unknown"
-    return cast(RunStartSource, value)
 
 
 def _try_import_sgxt() -> ModuleType | None:
