+++ framework/py/flwr/superlink/servicer/control/control_handlers.py
@@ -168,6 +168,7 @@ from flwr.superlink import extensions
 from flwr.superlink.artifact_provider import ArtifactProvider
 from flwr.superlink.auth_plugin import ControlAuthnPlugin
 from flwr.superlink.federation.noop_federation_manager import NoOpFederationManager
+from flwr.superlink.run_source import RunStartSource
 
 
 class InvalidConnectorRequestError(FlowerError):
@@ -468,7 +469,7 @@ def start_run(  # pylint: disable=too-many-branches,too-many-locals,too-many-sta
     state: LinkState,
     fleet_api_type: str | None,
     *,
-    source: extensions.RunStartSource = "unknown",
+    source: RunStartSource = "unknown",
 ) -> StartRunResponse:
     """Create run ID."""
     log(INFO, "ControlServicer.StartRun")
