+++ framework/py/flwr/superlink/servicer/control/control_servicer.py
@@ -98,9 +98,12 @@ from flwr.supercore.auth.typing import AccountInfo
 from flwr.supercore.error import ApiErrorCode, FlowerError
 from flwr.supercore.object_store import ObjectStoreFactory
 from flwr.supercore.utils import get_metadata_str
-from flwr.superlink import extensions
 from flwr.superlink.artifact_provider import ArtifactProvider
 from flwr.superlink.auth_plugin import ControlAuthnPlugin
+from flwr.superlink.run_source import (
+    RUN_SOURCE_METADATA_KEY,
+    resolve_run_start_source,
+)
 
 from . import control_handlers
 from .control_account_auth_interceptor import get_current_account_info
@@ -129,17 +132,17 @@ class ControlServicer(control_pb2_grpc.ControlServicer):
         self, request: StartRunRequest, context: grpc.ServicerContext
     ) -> StartRunResponse:
         """Create run ID."""
-        # Best-effort analytics attribution only; 
+        # Best-effort analytics attribution only;
         # trust caller-provided labels and default missing sources to unknown.
         metadata = context.invocation_metadata()
-        run_source = get_metadata_str(metadata, extensions.RUN_SOURCE_METADATA_KEY)
+        run_source = get_metadata_str(metadata, RUN_SOURCE_METADATA_KEY)
 
         return control_handlers.start_run(
             request,
             _get_account(),
             self.linkstate_factory.state(),
             self.fleet_api_type,
-            source=extensions.resolve_run_start_source(run_source),
+            source=resolve_run_start_source(run_source),
         )
 
     def StreamLogs(  # pylint: disable=C0103
